+++ src/test/java/hosh/testsupport/TemporaryFolder.java
@@ -23,7 +23,6 @@
  */
 package hosh.testsupport;
 
-import hosh.doc.Todo;
 import org.junit.jupiter.api.extension.AfterEachCallback;
 import org.junit.jupiter.api.extension.BeforeEachCallback;
 import org.junit.jupiter.api.extension.Extension;
@@ -36,7 +35,6 @@ import java.nio.file.Path;
 /**
  * Heavily inspired by JUnit 4.12 TemporaryFolder but with less features.
  */
-@Todo(description = "don't expose java.io.File, tests usually want java.nio.file.Path")
 public class TemporaryFolder implements Extension, BeforeEachCallback, AfterEachCallback {
 
 	private File folder;