+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 ### Changed
+- `walk`: terminate as soon as first error is raised
+
 ### Fixed
 
 ## [0.0.38] - 2020-02-29