+++ CHANGELOG.md
@@ -6,10 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2020-05-31
+
 ### Added
 - sign jar file with GPG when deploy on github
 
 ### Changed
+- introducing Java modules
 - `walk`: 'fail fast' error handling
 - removing automatic semicolon insertion